requirements.txt
Flask==3.0.3
peewee==3.17.6
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==21.2.0
Flask-SocketIO==5.3.6
python-socketio==5.11.3
eventlet==0.36.1
gunicorn
Flask-Cors==4.0.1
